SERVER-19500 Remove shard version checking for write command explain
Currently the write commands explain code calls
ensureShardVersionOKOrThrow, which performs legacy style shard version
checking based on the contents of the connection. This however does not
have any effect, because write commands are not sent over versioned
connections, so effectively write command explain is non-sharding aware
currently.

This change removes this check for the explain commands.